Fix flakiness in uao_crash_compaction_*
The flakiness was due to concurrent VACUUMs.  If there is another parallel drop
transaction (on any relation) active, the drop is skipped. This is to avoid
upgrade deadlock as the other drop transaction might be on the same relation.

We added additional test coverage for scenario of skipping the drop phase
during concurrent vacuum and crash recovery with file in drop pending state.

Co-authored-by: Ashwin Agrawal <aagrawal@pivotal.io>