Allow gpdbrestore to restore files from different dbids on datadomain
This was supposed to be supported with the addition of content id to
filenames, but it was not working.

Added feature to gpddboost --readFile option to accept regular
expression in filename and search the data domain directory for a file
that matches. If a file exists with the matching content id, but a
different dbid, we will find it and be able to restore.

Co-authored-by: Chris Hajas <chajas@pivotal.io>
Co-authored-by: Karen Huddleston <khuddleston@pivotal.io>