Readers should check abort status of their subxids
QE readers incorrectly return true for TransactionIdIsCurrentTransactionId()
when passed with an xid that is an aborted subtransaction of current
transaction.  The end effect is wrong results because tuples inserted by the
aborted subtransaction are seen (treated as visible according to MVCC rules) by
a reader.  Current patch fixes the bug by looking up abort status of an XID
from pg_clog.  In a QE writer, just like in upstream PostgreSQL, subtransaction
information is available in CurrentTransactionState (even when subxip cache has
overflown).  This information is not maintained in shared memory, making it
unavailable to a reader.  Readers must resort to a longer route to get the same
information - pg_subtrans and pg_clog.

The test case also needed a fix because the SQL query (insert into select *)
didn't result in a reader gang being created.  The SQL is changed to a join on
a non-distribution column so as to achieve reader gang creation.