Teach heap_truncate_one_rel to handle AO tables as well
Upstream commit cab9a0656 introduced an optimization to truncate tables
in scenarios that permit "unsafe" operations where we don't have to
churn on the relfilenode for the underlying tables. AO table got a free
ride but for the wrong reason.

This patch teaches heap_truncate_one_rel() to perform the unsafe /
optimal truncation on AO tables. This allows us to converge the callers
back to how they look like in Postgres 9.0.

Specifically, we're now able to inline TruncateRelfiles() back into
ExecuteTruncate() .

One caveat introduced by this patch though, is the "optimal" / unsafe
truncation of an AO table can potentially leak some disk space: we are
not performing a real file-level truncate, merely seeking back to offset
0 -- because the aoseg auxiliary table is truncated -- on the next
write, therefore the space after the EOF mark is wasted in some sense.

Co-authored-by: Jesse Zhang <sbjesse@gmail.com>