gptransfer supports implicit sequence and explicit sequence dependency
Because the pg_dump handle the implicit sequence (serial column type)
and explicit sequence differently in different gpdb version, so we
need to detect the sequences which not included in the table dump sql,
then we should create them firstly. And also for all dependent sequence,
setval() to source value after data transferred.

Signed-off-by: Ming Li <mli@pivotal.io>