behave: don't use demo-cluster for gpaddmirrors tests
The changes here are cherry picked from commit e9c66f27, but that commit
deals with WALrep logical changes we don't need. The main change here is
that we no longer pull up a demo cluster (the logic was broken); we
instead use the multi-host cluster set up by the CI. We additionally
delete the behave step that is no longer used.

Patch was originally
Co-authored-by: Nadeem Ghani <nghani@pivotal.io>
Co-authored-by: Jim Doty <jdoty@pivotal.io>