Provide more accurate debugging information for demo create failure.
There is case that a postmaster process is killed-by-9 and then the
process has no chance to remove its temporary files /tmp/.s.PGSQL.${PORT}*,
which are used by the demo create script to check the port usability.
For the above case re-creating demo cluster will fail but the error
message information as below is misleading since apparently the
port is not used.

  Check to see if the port is free by using :  'netstat -an | grep 16432'.