Do not look for an aborted parent of the given subxact
TransactionIdDidAbort() interface is designed to work for all transaction IDs.
It walks up the transaction hierarchy to look for an aborted parent if status
of the given transaction is found to be SUB_COMMITTED.  This is a wasted effort
when a QE reader wants to test if it's subtransaction has aborted or not.  This
commit introduces a new interface that avoids this wasted effort for QE
readers.

Also fix a unittest failure and add a scenario where parent of a transaction
subtree aborts while some children commit.