Move alert severity classification to callsite
Alert sending was previously triggered by invoking errSendAlert(true)
in the ereport() call. At the time of sending, the alerting code tried
to figure out how important/interesting the error message in question
was by string comparisons on the error message. This based was on a
set of strings which were hard to keep in sync with the codebase, and
rather suboptimal since the callsite where the error originated knows
how severe the error in question is.

Convert errSendAlert() to take a severity level instead such that the
callsite can instruct the alert sending code with how to handle the
alert. All ereports that generated an alert have been updated to match
the previous severity.

In the process, clean up style in a few ereport() calls and the alert
code in general.