Fix postjoin deduplication handling
With the merge, in convert_IN_to_join InClauseInfo->lefthand information
was included in the InClauseInfo, which impacted the implementation of
post join deduplication in case of semi join queries.

InClauseInfo was used in add_IN_vars_to_tlists to populate the
targetlist of the resulting relation (of the subquery).  The
InClauseInfo is also used to determine what attributes are needed for
which relids, maintained in attr_needed. attr_needed is used when
creating the target list of the join rel in build_joinrel_tlist.

Using this information, dedup information was attached to the rel in
`cdb_make_rel_dedup_info`, however the calculation to decide post join
or prejoin depends on the targetlist for the relation and due to the
inclusion of the lefthand relid in InClauseInfo, the targetlist of join
rel include the projection from the in clause which is not expected. Due
to it the subquery deduplication cannot be completed and planner failed
to create a plan. Consider the below setup.
```sql
CREATE TABLE dedup_test1 ( a int, b int ) DISTRIBUTED BY (a);
CREATE TABLE dedup_test2 ( e int, f int ) DISTRIBUTED BY (e);
CREATE TABLE dedup_test3 ( a int, b int, c int) DISTRIBUTED BY (a) PARTITION BY RANGE(c) (START(1) END(2) EVERY(1));
EXPLAIN SELECT * FROM dedup_test1 INNER JOIN dedup_test2 ON dedup_test1.a= dedup_test2.e WHERE (a) IN (SELECT a FROM dedup_test3);
ERROR:  Unexpected internal error (planmain.c:276)
```

This patch fixes the above issue and removes lefthand relid from the
InClauseInfo as it is not required to decide if post join dedup is
required on the In clause of the query. For the above setup, the join
between t1, t2, and t3 should not project the column of t3, also
lefthand relids are only populated in convert_IN_to_join and not used
anywhere, so removed the code all together.

Signed-off-by: Dhanashree Kashid <dkashid@pivotal.io>