ci: nightly-trigger only on AIX jobs
For a while there were several jobs that were behind the nightly
trigger.  This necessitated some logic about including the
nightly-trigger resource if any of a number of conditions were met.  At
the time of this commit, the only job that is using the resourse is an
AIX job.  Therefore the inclusion of the nightly-trigger resource will
match the conditions that include the only job that requires that
resource.  This elimnates the "resource not used" error that can be seen
when setting a development version of the pipeline that does not include
the AIX job.

Authored-by: Jim Doty <jdoty@pivotal.io>