Remove rd_issyscat from RelationData.
Greenplum added rd_isyscat to Relation structure. Only usage of the same is in
markDirty() to decide if buffer should be marked dirty or not. The setting of
rd_issyscat is based in checking if relation name starts with "pg_" then set it
else not. Which anyways is very loose.

Modified instead to just make check based on if oid < FirstNormalObjectId or to
cover for pg_aoseg tables RelationGetNamespace(relation) ==
PG_AOSEGMENT_NAMESPACE. So, this allows us to remove the extra variable.

This patch is not trying to change the intent of GUC
`gp_disable_tuple_hints`. That's all together different discussion.