Remove default statistics control in join planning
Commit e0409357fe63a2 moved to using default estimates rather
than interrogating the QEs for relations which lack statistics.
As an effect of this, the cdb_default_stats_used member was
hardcoded to false and the warnings for missing statistics did
never fire.

Rather than resurrecting the warnings, this removes the code
that attempts to figure out if the warnings at all apply since
it seems quite expensive to run that in the hot path of every
join query being planned.