Remove FIXME from ao_insert_replay().
The AO implementation aligns with 8.4 and forward heap implementation, to write
the data during recovery and not fail.

Also, to note in case of AO the way seek it performed during replay, its not
going to fail if file doesn't have that much data yet. As the way seek works
irrespective of the length of file will seek to that offset from requested
position and write the data (sure file will have hole in it for this case) . But
it will not result in seek failure as such. We will write the data and if
truncation has happened then will happen again during recovery.