Remove dead test code
The ability to inject a pause for the FTS prober was added in 2011
in order to test the resolution to a bug in the FTS prober.  There
is no test that currently use this, and seemingly there never was.
Further, the comment is incorrect which adds to the suspicion that
this has been unused for quite some time so remove.