Disable concurrent index build entirely
The patch removes gp_create_index_concurrently GUC.  There is no use setting
this GUC, concurrent index building is not supported in Greenplum.  Greenplum
needs additional work for this feature to behave properly.  Also undone is an
incorrect (and dangerous) attempt to make this feature work in Greenplum.  The
change was dangerous because IndexStmt was dispatched to QEs outside of a
distributed transaction/2PC, which is a recipe for inconsistency.

GitHub issue #5293 is created to track the work needed for this feature.