Remove guc gp_setwith_alter_storage.
Alter table change storage type feature is hidden under this guc.  The syntax
ALTER TABLE <tablename> SET WITH (appendonly=true/false) is not documented as
well, since seems this feature is not full ready for primetime. Hence remove the
guc and keep it disabled, anytime in future can code the feature fully and
expose the same.

Github issue #5300 tracks to full enable the feature.