Unbreak compilation of SNMP support
Commit 3b3adb2be3c561fd7638bed96b859ee2e408f3a2 most likely broke
compilation of the SNMP related code, as it required feature test
macros to enable <sys/types.h>.  There however seems to be little
reason for the feature test macros here either, so remove and try
to unbreak the build.  Also enable snmp in Travis CI so we get a
heads up the next time we break it.