Temporarily minimize the pg_upgrade test dataset
The regression database has shown to be full of questionable
objects which we probably won't support for upgrades without
manual intervention before the upgrade. With all the changes
going into pg_upgrade it's time to start on a known minimal
dataset and expand from there to the full ICW leftover schema.
As a stop-gap for now, drop the regression databases eand only
upgrade a small subset.

Also disallow upgrades of orhpaned toast tables, and clean
up the invocation of the test_gpdb_pre.sql script to be less
hacky.