Plug some trivial memory leaks in pg_dump and pg_upgrade.
This is a partial backport of the below commit from upstream, with
one hunk removed as it touches code yet in the future of this fork
and another hunk massaged to account for path changes. Since we
want the pg_upgrade hunk from this commit to silenve Coverity, we
may as well grab all the hunks that are relevant and close those
leaks ahead time in the fork.

  commit f712289ffad7c3fb6eb3be4f81adb0aa0981c9f7
  Author: Heikki Linnakangas <heikki.linnakangas@iki.fi>
  Date:   Thu Jul 2 20:58:51 2015 +0300

    Plug some trivial memory leaks in pg_dump and pg_upgrade.

    There's no point in trying to free every small allocation in these
    programs that are used in a one-shot fashion, but these ones seems like
    an improvement on readability grounds.

    Michael Paquier, per Coverity report.