Fix a few small leaks in pg_upgrade
If we don't find any AO tables we exit early, but we failed to close
the PQExpBuffer we had for the query. Fix by destroying the buffer
explicitly.

Move the freeing of numeric_rels to be unconditionally since pg_free
can cope with a NULL pointer.

Save the quote_identifier() returned strings in a char * rather than
passing them to fprint() so we can pg_free() them on the way out.