Fix partitionrank query for external partitions
When dumping a partitioning hierarchy with external members, pg_dump
will create the external table separately and EXCHANGE it into the
hierarchy since CREATE TABLE .. PARTITION BY cannot express external
tables. The query for finding the rank at which to EXCHANGE the table
into had a bug which made the external partition end up at a different
place in the hiearchy on restoring.

	create table t (i integer, j integer)
		distributed by (i)
		partition by range (j) (start(1) end(3) every(1));
	create readable external table tt_external (i int, j int)
		location ('gpfdist://host.invalid:8000/file') format 'text';
	alter table t exchange partition
		for (rank(2)) with table tt_external without validation;

This yielded the following DDL in the dump (edited only for line
width):

	CREATE TABLE t (
	    i integer,
	    j integer
	) DISTRIBUTED BY (i) PARTITION BY RANGE(j)
	          (
	          START (1) END (2) EVERY (1) WITH (tablename='t_1_prt_1',
			  									appendonly=false ),
	          START (2) END (3) EVERY (1) WITH (tablename='t_1_prt_2',
			  									appendonly=false )
	          );
	ALTER TABLE t EXCHANGE PARTITION
		FOR (RANK(1)) WITH TABLE t_1_prt_2_external_partition__
		WITHOUT VALIDATION;

This effectively reverts 2ea6aec39cffca1d9028ebacaf50e7eea5664854
which attempted to optimize the partitionrank query, but fell a
brick shy of the intended load it seems. Since external partitions
are quite rare, it hardly seems worth optimizing this case so going
back to querying the base view which is known to be correct.