binary_swap: drop public.toast_chunk_test before testing
The regression table for the TOAST_MAX_CHUNK_SIZE changes can't be
tested as part of binary-swap, because the prior version of the database
doesn't have the bugfix necessary to correctly dump that table.

Add a pretest_cleanup.sql which will handle these cases. We simply drop
the pieces of the database that can't be correctly tested.

Co-authored-by: Shoaib Lari <slari@pivotal.io>