Fix redistribute bug on some types which need to convert
After 8.4 merge, we have two restrictlist 'mergeclause_list'
and 'hashclause_list' in function 'add_paths_to_joinrel'. We
use mergeclause_list in cdb motion in hashjoin. But some of
keys should not been used as distribution keys.

We can use restrictinfo->hashjoinoperator to determine if
a RestrictInfo in mergeclause_list can be a redistribute
key. Considered there are different requirements for
RestrictInfo for redistribution and merge, We can make a
new function to make a cdbclause_list,rather than use
mergeclause_list as redistribute keys in merge and hash join.