Avoids deadlocks for concurrent index creation on AO tables.
Creating the block directory auxiliary table for an append-only table requires
an AccessExclusive lock (we're not sure why). We defer the creation of the
block directory until creation of the first index on an AO table. This leads to
a lock upgrade later, which causes a dead lock when two concurrent sessions
both believe they are creating the first index.

Acquire a ShareRowExclusiveLock on the target table while creating an index.
This allows only one of the concurrent users to have access to the table at a
time, and then allows the winner to aquire an AccessExclusive lock when
creating the AOBlkDir table.

Co-authored-by: Jimmy Yih <jyih@pivotal.io>
Co-authored-by: Jesse Zhang <sbjesse@gmail.com>