Remove a fixme in make_result_inplace in numeric.c
Before we merge postgres91, Numeric just have one type. But after
that, we have NumericLong and NumericShort. We can make a short
numeric when the data is mall, and make a long numeric when the
data is large.
The variable is a sum in NumericAvgTransData. So it won't
save a lot of space, Just in line with upstream.