Avoid using pg_locks with distributed tables
In upstream Postgres, pg_locks exposes part of the lock manager so that
DBA's can inspect locks taken by various backends. In Greenplum, we
modified pg_lock_status() -- the function that underlies pg_locks -- to

  a) provide additional Greenplum-specific information (e.g.
     mppsessionid); and

  b) aggregate the locks from master and all primary segments.

One consequence of the implementation we chose to achieve point b above,
is that queries that involve both pg_locks and a distributed table won't
work. If you're lucky (planner or ORCA places the function call in the
top slice), it "won't work" and throws an error like this at you:

ERROR:  query plan with multiple segworker groups is not supported

If you're not lucky (planner or ORCA schedules the function call on a
different slice running on master), it most likely secretly doesn't do
point b and only returns locks from the master.

Before we fix pg_locks, rewrite the isolation2 test case "starve_case"
to separate the repeated queries in pg_locks from the main query. This
does logically the same thing in PL/pgSQL but it's now safe.

This commit also adds starve_case back to the isolation2 schedule.