Reimplement DISTINCT for window aggregates.
GPDB 5 supported DISTINCT in window aggregates, e.g:

COUNT(DISTINCT x) OVER (PARTITION BY y)

However, PostgreSQL does not support that, and as a result, GPDB lost that
capability as part of the window functions rewrite, too. In the upstream,
there's an explicit check for that, that it was lost in the window function
rewrite. So the parser accepted that, but it was executed just like if
there was no DISTINCT. There were also no tests for this, that would return
a different result with the DISTINCT than without, which is why no-one
noticed it.

To fix, implement the DISTINCT support, to the same extent that the old
implementation supported it. The new implementation adds a little sort +
deduplicate step for each DISTINCT aggregate. I'm not sure how this
compares with the old implementation, performance-wise, but at least it
works now.

Also, add the missing tests.