Disable DISTINCT-qualified aggregates with ORCA.
The new regression tests revealed that it doesn't work. With an assertion-
enabled ORCA build, I got an assertion failure like this:

2018-08-23 11:20:08:371479 EEST,THD000,ERROR,""/home/heikki/gpdb/optimizer-main/libgpos/include/gpos/common/CDynamicPtrArray.h:300: Failed assertion: pos < m_size && ""Out of bounds access""
Stack trace:
1    0x00007f363fb3e78a gpos::CException::Raise + 252
2    0x00007f3640be0970 gpos::CDynamicPtrArray + 84
3    0x00007f3640c93dac gpopt::CWindowPreprocessor::SplitPrjList + 1162
4    0x00007f3640c9404b gpopt::CWindowPreprocessor::SplitSeqPrj + 303
5    0x00007f3640c94b61 gpopt::CWindowPreprocessor::PexprSeqPrj2Join + 357
6    0x00007f3640c95276 gpopt::CWindowPreprocessor::PexprPreprocess + 316
7    0x00007f3640c240a2 gpopt::CExpressionPreprocessor::PexprPreprocess + 1098
8    0x00007f3640bc2d62 gpopt::CQueryContext::CQueryContext + 696
9    0x00007f3640bc36df gpopt::CQueryContext::PqcGenerate + 1413
10   0x00007f3640c95d86 gpopt::COptimizer::PdxlnOptimize + 1042
11   0x000055b2e8252f26 COptTasks::OptimizeTask + 1488
12   0x00007f363fb58a0d gpos::CTask::Execute + 183
13   0x00007f363fb5d447 gpos::CWorker::Execute + 199
14   0x00007f363fb56d77 gpos::CAutoTaskProxy::Execute + 287
15   0x00007f363fb3479b gpos_exec + 800
"",",,,,,,"explain  select dt, pn, sum(distinct pn) over (partition by dt), sum(pn) over (partition by dt) from sale;",0,,"COptTasks.cpp",545,
2018-08-23 11:20:08.372392 EEST,"heikki","postgres",p19807,th1163394560,"[local]",,2018-08-23 11:19:53 EEST,0,con4,cmd7,seg-1,,dx6,,sx1,"LOG","00000","Planner produced plan :1",,,,,,"explain  select dt, pn, sum(distinct pn) over (partition by dt), sum(pn) over (partition by dt) from sale;",0,,"orca.c",61,

This caused the query to fall back to planner, which worked. But with
assertions disabled, it crashed instead.

We should fix ORCA to deal with that. One option is to rip out all the
special code to plan DISTINCT-qualified aggregates in ORCA, and just pass
through the windistinct flag to the executor. That's basically what the
Postgres planner does, and the executor will deal with deduplicating the
input. But for now, let's just stop the crashing.